+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/StubExoPlayer.java
@@ -398,12 +398,7 @@ public class StubExoPlayer extends StubPlayer implements ExoPlayer {
   }
 
   @Override
-  public void experimentalSetOffloadSchedulingEnabled(boolean offloadSchedulingEnabled) {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public boolean experimentalIsSleepingForOffload() {
+  public boolean isSleepingForOffload() {
     throw new UnsupportedOperationException();
   }